Little refactor on data layer
