+++ app/src/main/java/com/rafaels/warhammertowbuilder/di/AppModules.kt
@@ -4,7 +4,7 @@ import com.rafaels.data.RestClient
 import com.rafaels.data.error.ErrorHandler
 import com.rafaels.data.repository.MockRepositoryImpl
 import com.rafaels.data.repository.UnitRepositoryImpl
-import com.rafaels.domain.UnitRepository
+import com.rafaels.domain.repository.UnitRepository
 import com.rafaels.domain.usecase.GetUnit
 import com.rafaels.warhammertowbuilder.WarhammerTOWBuilderApplication
 import com.rafaels.warhammertowbuilder.ui.MainViewModel
