+++ data/src/main/java/com/rafaels/data/RestClient.kt
@@ -3,23 +3,11 @@ package com.rafaels.data
 import com.rafaels.data.api.UnitApi
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
+import okhttp3.logging.HttpLoggingInterceptor.*
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
 object RestClient {
-
-    private fun createOkHttpClient(): OkHttpClient {
-        return OkHttpClient.Builder()
-            .addInterceptor(
-                HttpLoggingInterceptor().setLevel(
-                    if (BuildConfig.DEBUG)
-                        HttpLoggingInterceptor.Level.BODY
-                    else
-                        HttpLoggingInterceptor.Level.NONE
-                )
-            )
-            .build()
-    }
 
     private val retrofitInstance: Retrofit = Retrofit.Builder()
         .baseUrl(BuildConfig.API_URL)
@@ -28,4 +16,15 @@ object RestClient {
         .build()
 
     fun getUnitApi(): UnitApi = retrofitInstance.create(UnitApi::class.java)
+
+    private fun createOkHttpClient(): OkHttpClient {
+        return OkHttpClient.Builder()
+            .addInterceptor(getLoggingInterceptor())
+            .build()
+    }
+
+    private fun getLoggingInterceptor() = HttpLoggingInterceptor()
+        .setLevel(getLoggingInterceptorLevel())
+
+    private fun getLoggingInterceptorLevel() = if (BuildConfig.DEBUG) Level.BODY else Level.NONE
 }
