+++ domain/src/main/java/com/rafaels/domain/repository/UnitRepository.kt
@@ -1,5 +1,6 @@
-package com.rafaels.domain
+package com.rafaels.domain.repository
 
+import com.rafaels.domain.Resource
 import com.rafaels.domain.model.UnitModels
 
 interface UnitRepository {
