+++ domain/src/main/java/com/rafaels/domain/usecase/GetUnit.kt
@@ -1,7 +1,7 @@
 package com.rafaels.domain.usecase
 
 import com.rafaels.domain.Resource
-import com.rafaels.domain.UnitRepository
+import com.rafaels.domain.repository.UnitRepository
 import com.rafaels.domain.model.UnitModels
 
 class GetUnit(
